requirement.txt
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.0.4
Babel==2.6.0
beautifulsoup4==4.6.3
bleach==2.1.3
blinker==1.4
Brlapi==0.6.7
bs4==0.0.1
buildozer==0.34
certifi==2018.8.24
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.0
command-not-found==0.3
cryptography==2.3
cupshelpers==1.0
cycler==0.10.0
Cython==0.25.2
decorator==4.3.0
defer==1.0.6
distro==1.3.0
distro-info==0.18
docutils==0.14
entrypoints==0.2.3.post3
future==0.16.0
geocoder==1.38.1
googletrans==2.3.0
gTTS==2.0.1
gTTS-token==1.1.2
Guake==3.3.0
html5lib==1.0.1
httplib2==0.11.3
icalendar==4.0.0
idna==2.6
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.4
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==15.1.0
keyrings.alt==3.1
Kivy==1.10.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.6.0
lxml==4.2.5
macaroonbakery==1.1.4
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
mathparse==0.1.2
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
nltk==3.3
notebook==5.4.1
numpy==1.14.5
oauth==1.0.1
oauthlib==2.0.6
olefile==0.46
pandas==0.23.3
pandocfilters==1.4.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==2.0.6
protobuf==3.0.0
psutil==5.4.6
PyAudio==0.2.11
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.4
Pygments==2.2.0
PyGObject==3.30.1
PyJWT==1.6.4
pylint==2.1.1
pymacaroons==0.13.0
pymongo==3.7.2
PyNaCl==1.2.1
pyparsing==2.2.2
pyRFC3339==1.1
python-apt==1.7.0
python-dateutil==2.6.1
python-debian==0.1.33
python-for-android==0.6.0
python-twitter==3.4.2
python-xlib==0.23
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.0
qtconsole==4.4.2
ratelim==0.1.6
remarkable==1.87
reportlab==3.5.6
requests==2.18.4
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
safeeyes==2.0.6
scikit-learn==0.19.2
scipy==1.1.0
SecretStorage==2.3.1
Send2Trash==1.4.2
sh==1.12.4
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
sklearn==0.0
SpeechRecognition==3.8.1
SQLAlchemy==1.2.12
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.3
wcwidth==0.1.7
weather-api==1.0.6
webencodings==0.5
widgetsnbextension==2.0.1
wrapt==1.10.11
xkit==0.0.0
zope.interface==4.3.2
